+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # :package: 0.1.0
 
 01. :gift: Added interface `dev.orne.beans.Identity`
+01. :gift: Added interface `dev.orne.beans.IdentityBean`
+01. :gift: Added annotation `dev.orne.beans.BeanReference`
+01. :gift: Added annotation `dev.orne.beans.ValidBeanReference`
 01. :gift: Added class `dev.orne.beans.IdentityToken`
 01. :gift: Added class `dev.orne.beans.BeanValidationUtils`
 01. :gift: Added class `dev.orne.beans.BeanAnnotationFinder`
-01. :gift: Added annotation `dev.orne.beans.BeanReference`
-01. :gift: Added annotation `dev.orne.beans.ValidBeanReference`
 01. :gift: Added Apache `beanutils` converters
     01. :gift: Added class `dev.orne.beans.converters.OrneBeansConverters`
     01. :gift: Added class `dev.orne.beans.converters.EnumConverter`